TMXLayer now also supports scaling.

+++ src/org/anddev/andengine/util/MathUtils.java
@@ -162,8 +162,8 @@ public class MathUtils implements MathConstants {
 		return revertRotateAroundCenter(pVertices, pRotation, pRotationCenterX, pRotationCenterY);
 	}
 
-	public static int bringToBounds(final int pMinValue, final int pMaxValue, final float pValue) {
-		return (int)Math.max(pMinValue, Math.min(pMaxValue, pValue));
+	public static int bringToBounds(final int pMinValue, final int pMaxValue, final int pValue) {
+		return Math.max(pMinValue, Math.min(pMaxValue, pValue));
 	}
 
 	// ===========================================================
